Simplifying dfca

+++ deformetrica/in_out/image_functions.py
@@ -1,8 +1,5 @@
 import os.path
 import sys
-
-from ..core import default
-
 sys.path.append(os.path.dirname(os.path.abspath(__file__)) + os.path.sep + '../../../')
 
 import numpy as np
@@ -12,9 +9,9 @@ def normalize_image_intensities(intensities):
 
     dtype = str(intensities.dtype)
     if dtype == 'uint8':
-        return np.array(intensities / 255.0, dtype=default.dtype), dtype
+        return np.array(intensities / 255.0, dtype="float32"), dtype
     else:
-        return np.array(intensities, dtype=default.dtype), dtype
+        return np.array(intensities, dtype="float32"), dtype
 
 
 def rescale_image_intensities(intensities, dtype):
